Server/requirements.txt
django==4.2.7
# psycopg2==2.9.9
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1